SCI: ROGER: Make sweep-SVG PNG size casts explicit
Review follow-up: the MemoryWriteStreamDynamic size() results are
int64 and narrowed silently to the builder's uint32 lengths; make
the casts explicit. Also correct a stale test-helper comment.

+++ engines/sci/roger/utils/studio/roger_sweep_svg.cpp
@@ -258,8 +258,8 @@ Common::String buildSweepSvg(const Graphics::Surface &left,
 		Common::MemoryWriteStreamDynamic pl(DisposeAfterUse::YES);
 		Common::MemoryWriteStreamDynamic pr(DisposeAfterUse::YES);
 		if (encodeSurfacePng(*ls, pl) && encodeSurfacePng(*rs, pr)) {
-			out = buildSweepSvgFromPngData(pl.getData(), pl.size(),
-			                               pr.getData(), pr.size(),
+			out = buildSweepSvgFromPngData(pl.getData(), static_cast<uint32>(pl.size()),
+			                               pr.getData(), static_cast<uint32>(pr.size()),
 			                               ls->w, ls->h, animated);
 		}
 	}
